requirements.txt
flask==2.3.3
gunicorn==21.2.0
pytest==7.4.0
python-dotenv==1.0.0
boto3==1.26.0
reportlab==3.6.12
pyjwt==2.6.0
cryptography==39.0.0
requests==2.28.2
